requirements.txt
affine==2.4.0
attrs==25.4.0
blessings==1.7
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cligj==0.7.2
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
earthengine-api==1.7.16
fonttools==4.61.1
geopandas==1.1.2
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
httplib2==0.31.2
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pathlib==1.0.1
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
rasterio==1.5.0
requests==2.32.5
rsa==4.9.1
scipy==1.17.1
shapely==2.1.2
six==1.17.0
uritemplate==4.2.0
urllib3==2.6.3
